Extract radio selection into web/js/ui/radio-catalog.js
The make/model dropdowns, the search box with its autocomplete list, the
last-radio cookie, and the CHIRP metadata load that follows a selection now
live in one module owning the search-result state. Its own listeners bind
from the module.
